# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hola Municipality Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Deactivate department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # UNITS
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores
      responses:
        "200":
          description: Stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Deactivate store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # INVENTORY ITEMS
  /inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: category
          in: query
          schema:
            type: ["number", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Soft-delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # PURCHASE ORDERS (LPO)
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
...
    Department:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    # UNIT
...
    Store:
      type: object
      required: [id, storeCode, name, isActive, createdAt]
      properties:
        id:
          type: number
        storeCode:
          type: string
        name:
          type: string
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    StoreInput:
      type: object
      required: [storeCode, name]
      properties:
        storeCode:
          type: string
        name:
          type: string
        location:
          type: string
        description:
          type: string
    StoreUpdate:
      type: object
      properties:
        storeCode:
          type: string
        name:
          type: string
        location:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    # INVENTORY ITEM
    InventoryItem:
      type: object
      required: [id, itemCode, itemName, categoryId, unitOfMeasure, currentQuantity, minimumStock, maximumStock, reorderLevel, purchasePrice, status, createdAt]
      properties:
        id:
          type: number
        itemCode:
          type: string
        barcodeQr:
          type: ["string", "null"]
        itemName:
          type: string
        categoryId:
          type: number
        categoryName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        unitOfMeasure:
          type: string
        currentQuantity:
          type: number
        minimumStock:
          type: number
        maximumStock:
          type: number
        reorderLevel:
          type: number
        shelfBinLocation:
          type: ["string", "null"]
        purchasePrice:
          type: number
        supplierId:
          type: ["number", "null"]
        supplierName:
          type: ["string", "null"]
        dateReceived:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, low_stock, out_of_stock]
        createdAt:
          type: string
    InventoryItemInput:
      type: object
      required: [itemName, categoryId, unitOfMeasure, currentQuantity, minimumStock, maximumStock, reorderLevel, purchasePrice]
      properties:
        itemCode:
          type: string
        barcodeQr:
          type: string
        itemName:
          type: string
        categoryId:
          type: number
        description:
          type: string
        unitOfMeasure:
          type: string
        currentQuantity:
          type: number
        minimumStock:
          type: number
        maximumStock:
          type: number
        reorderLevel:
          type: number
        shelfBinLocation:
          type: string
        purchasePrice:
          type: number
        supplierId:
          type: number
        dateReceived:
          type: string
        expiryDate:
          type: string
    InventoryItemUpdate:
      type: object
      properties:
        itemCode:
          type: string
        barcodeQr:
          type: string
        itemName:
          type: string
        categoryId:
          type: number
        description:
          type: string
        unitOfMeasure:
          type: string
        currentQuantity:
          type: number
        minimumStock:
          type: number
        maximumStock:
          type: number
        reorderLevel:
          type: number
        shelfBinLocation:
          type: string
        purchasePrice:
          type: number
        supplierId:
          type: number
        dateReceived:
          type: string
        expiryDate:
          type: string

    # PURCHASE ORDER
...