# lib/api-zod/src/generated/types/departmentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hola Municipality Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentUpdate {
  name?: string;
  description?: string;
  isActive?: boolean;
}