# lib/api-zod/src/generated/types/inventoryItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hola Municipality Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemInput {
  itemCode?: string;
  barcodeQr?: string;
  itemName: string;
  categoryId: number;
  description?: string;
  unitOfMeasure: string;
  currentQuantity: number;
  minimumStock: number;
  maximumStock: number;
  reorderLevel: number;
  shelfBinLocation?: string;
  purchasePrice: number;
  supplierId?: number;
  dateReceived?: string;
  expiryDate?: string;
}