# lib/api-zod/src/generated/types/store.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hola Municipality Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Store {
  id: number;
  storeCode: string;
  name: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt?: string;
}