# lib/api-zod/src/generated/types/storeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hola Municipality Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreInput {
  storeCode: string;
  name: string;
  location?: string;
  description?: string;
}